requirements.txt
Django==4.1
django-extensions==3.2.1
django-import-export==2.9.0
djangorestframework==3.14.0
drf-nested-routers==0.93.4
psycopg2-binary==2.9.3